refactor: memperbaiki struktur file
