+++ app/src/main/java/com/example/pustakaku/components/BottomNavigation.kt
@@ -1,4 +1,4 @@
-package com.example.pustakaku
+package com.example.pustakaku.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -21,6 +21,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.unit.dp
+import com.example.pustakaku.R
 
 class BottomNavigation {
 }
