+++ app/src/main/java/com/example/pustakaku/core/navigation/Navigation.kt
@@ -10,7 +10,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Home
-import androidx.compose.material.icons.filled.List
 import androidx.compose.material.icons.filled.Menu
 import androidx.compose.material.icons.filled.Person
 import androidx.compose.material3.Surface
@@ -20,8 +19,8 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.platform.LocalContext
 import androidx.navigation.NavType
 import androidx.navigation.navArgument
-import com.example.pustakaku.BottomMenu
-import com.example.pustakaku.BottomMenuContent
+import com.example.pustakaku.components.BottomMenu
+import com.example.pustakaku.components.BottomMenuContent
 import com.example.pustakaku.features.homepage.HomeScreen
 //import com.example.pustakaku.SplashScreen
 import com.example.pustakaku.features.auth.data.repository.AuthRepository
@@ -37,11 +36,7 @@ import com.example.pustakaku.features.chapter_content.ReadingBookScreen
 //import com.example.pustakaku.features.auth.presentation.ui.viewmodel.AuthView
 import com.example.pustakaku.features.detail_book.BookDetailScreen
 import com.example.pustakaku.features.history.HistoryPage
-import com.example.pustakaku.features.profile.data.repository.ProfileRepository
-import com.example.pustakaku.features.profile.domain.usecase.GetUserUseCase
-import com.example.pustakaku.features.profile.domain.usecase.LogoutUseCase
 import com.example.pustakaku.features.profile.presentation.ProfileScreen
-import com.example.pustakaku.features.profile.presentation.ProfileViewModel
 
 @Composable
 fun Navigation() {
